[refactor] #81 navigation 패키지 생성
