+++ app/src/main/java/org/android/bbangzip/presentation/ui/subject/addsubject/navigation/AddSubjectNavigation.kt
@@ -1,9 +1,10 @@
-package org.android.bbangzip.presentation.ui.subject.addsubject
+package org.android.bbangzip.presentation.ui.subject.addsubject.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import kotlinx.serialization.Serializable
+import org.android.bbangzip.presentation.ui.subject.addsubject.AddSubjectRoute
 
 fun NavController.navigateToAddSubject() {
     navigate(
